fix (python): 整理python项目结构，可直接运行test中脚本
