+++ operatorspy/__init__.py
@@ -1,3 +1,7 @@
-from .liboperators import *
+import os
+import sys
+sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '.')))
+from .liboperators import open_lib, to_tensor, Kernel, Operator, ConstTensor, MutableTensor
 from .operators import OptypeEnum
 from .devices import DeviceEnum
+from .data_layout import *
